Tests/LibWeb/Text/input/css/border-image-non-fill-keyword.html

<!DOCTYPE html>
<script src="../include.js"></script>
<style>
    #foo {
        border-image: 100% none;
    }
</style>
<script>
    test(() => {
        println(document.styleSheets[0].cssRules[0].cssText);
    });
</script>
