Tests/LibWeb/Text/expected/css/border-image-non-fill-keyword.txt

#foo { border-image: none 100% / 1 / 0 stretch; }
